Restore User model
If CERN login is used, maybe we are going to have user who access to CERN applications
via University email, their access would be denied with this change. Related to User primary ID,
CERN SSO ensures the primary alias is always set in the token for CERN users. Administrators should be
aware of not using the "login email address" pattern to register permissions to a User, they should use the primary email alias instead